Patch from Wayne <dnrd@useunix.net>
--
We have found that with our rather old and antiquated HPUX boxes the
resolver may reuse the same query IDs frequently.  This was causing dnrd
to throw out the new query and reply with the wrong response as it
assumes query uniqueness based on the clients query ID.

This was fixed by examining not only the query ID but also the client's
source address and source port. It may even be a good idea to keep and
copy of the original query and memcmp() when all the other attributes
match.
--
We were seeing several warnings about bind() errors attributed to a port
already in use when trying to allocate a new random port.

I solved this by simply wrapping random port allocator with a while loop
to handle the EADDRINUSE case.
--
We use a shell script that runs processes in a foreground loop, waits
for them to terminate and restarts them. We usually want to keep
processes in the foreground but prevent unnecessary debug information
on standard our/error.

I hacked in a debug level of 255 which prevents dnrd from forking and
then switches the debug level to 0 to prevent debug messages.