Patch from Jarrod Lowe:
The attached patch scans a DNS reply received from an upstream DNS
server, and selects the lowest TTL in it. It then uses this as the
expiry time for the packet, rather then the value of CACHE_TIME. It
will also force-expire an old packet if you try to access it from the
cache.